+++ src/activities/ActivityManager.h
@@ -90,6 +90,7 @@ class ActivityManager {
 
   // goTo... functions are convenient wrapper for replaceActivity()
   void goToFileTransfer(std::string returnBookPath = {});
+  void goToNearbyStatsSync();
   void goToSettings();
   void goToFileBrowser(std::string path = {});
   void goToRecentBooks();